Create README.md
Today's embedded systems are highly heterogeneous in terms of technologies, security and protocols.
This raises the problem of harmonizing security policies. 
So it's essential to be equipped to detect inconsistencies between policies. However, formal expressions of security policies are needed to formally verify their compatibility or otherwise.  Hence my interest in formal languages and methods. I'm currently exploring the flexibility of the REGO language.